+++ application/client.core/src/app/environment/components/views/measurement/charts/styles.less
@@ -1,0 +1,20 @@
+@import '../../../../theme/variables.less';
+
+:host {
+    position: absolute;
+    display: block;
+    padding: 0;
+    margin: 0;
+    bottom: 0;
+    right: 0;
+    left:0;
+    top:0;
+    overflow-x: hidden;
+    overflow-y: auto;
+    & div.wrapper{
+        position: absolute;
+        top:0;
+        bottom: 0;
+        width: 100%;
+    }
+}
