+++ application/client.core/src/app/environment/components/views/measurement/charts/template.html
@@ -1,0 +1,3 @@
+<div class="wrapper" [ngStyle]="{'height': _ng_getChartsHeight()}" (contextmenu)="_ng_onChartContexMenu($event)">
+    <canvas #canvas></canvas>
+</div>
