+++ application/client.core/src/app/environment/components/views/measurement/overview/styles.less
@@ -1,0 +1,13 @@
+@import '../../../../theme/variables.less';
+
+:host {
+    position: absolute;
+    display: block;
+    padding: 0;
+    margin: 0;
+    bottom: 0;
+    right: 0;
+    left:0;
+    top:0;
+    overflow: hidden;
+}
