+++ application/client.core/src/app/environment/components/views/measurement/overview/template.html
@@ -1,0 +1,1 @@
+<canvas #canvas></canvas>
